+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.1
+
+* Add create ~ resolution database field
+  * (start=item["create_date"], end=item["resolution_date"])
+
+* Change property type
+  * issue_type, project_name, priority, status
+
+* Add create/update date string in page body
+
 ## 1.0
 
 init release